+++ src/App.tsx
@@ -7,11 +7,11 @@ import {Sidebar} from './components/Sidebar';
 import {WelcomeScreen} from './components/WelcomeScreen/WelcomeScreen.tsx';
 import {Cog6ToothIcon} from "@heroicons/react/24/outline";
 import {NotificationProvider} from './context/NotificationContext';
-import {NotificationContainer} from "./components/NotificationContainer.tsx";
 import {Auth0Provider} from '@auth0/auth0-react';
 import AIIcon from "./components/AIIcon.tsx";
 import {ChatContainer} from "./components/ChatContainer/ChatContainer.tsx";
 import {ModelControls} from "./components/ModelControls/ModelControls.tsx";
+import {NotificationContainer} from "./components/Notification/NotificationContainer.tsx";
 
 interface ModelSettings {
     temperature: number;
